+++ src/main/java/org/twuni/money/exchange/anet/command/PayCommand.java
@@ -18,7 +18,7 @@ public class PayCommand extends Command {
 		super( client, url );
 	}
 
-	public String execute( String accountNumber, String expirationDate, float amount, String relayUrl, long invoiceNumber, String notes ) throws IOException {
+	public String execute( String accountNumber, String expirationDate, double amount, String relayUrl, long invoiceNumber, String notes ) throws IOException {
 
 		List<NameValuePair> parameters = new ArrayList<NameValuePair>();
